syncbase: syncQL: support k not like "<wildcard-string>"
As pointed out by Ivan, since k(ey) <> "<string>" expressions
are now supported, there isn't a good reason to exclude k not
like expressions.

This change adds the support.  Every k/v pair except the range
implied by the wildcard expression will be considered for
inclusion.  Also, if the wildcard string contains no wildcards,
the expression is optimized to a k <> expression.

There are two bugfixes included in this change.  There was a bug
in the resultstream implementation of the test in query/test.
There was also a bug in the actual syncbase implementation of
the same resultstream.

Change-Id: If44650509fc8f570c58f55424cba524c9dcd1220