ref/services/groups: requiring a build tag to build the leveldb implementation
This is an attempt to avoid requiring leveldb for building and testing
the Vanadium core. The hope is that hiding the leveldb dependencies
behind a build tag will prevent "go get" from fetching the syncbase
repository.

This CL also aligns the command-line flags used by the groups services
with those used by the syncbase service.

Change-Id: I7b4e1b9575f29fc510fefc16d191c1f161d04bd0