v.io/syncbase/x/ref/services/syncbase/localblobstore/fs_cablobstore: Bug fix: blob reader code reading at EOF
I had a bug in the blob reader code triggered by performing a read at the end of file mark.
It would return an error indicating an illegal seek position, rather than simmply returning EOF.

This change returns EOF, as expected.

It also introduces a test that would have caught this.
It emnulates part of what syncbase does, by creating en empty blob,
and the using ResumeBlobWriter() to extend it.  This process happens to
read the empty blob, and hence hist the bug.

The test is called from the tests of both the fs_cablobstore (implementation
layer) and localblobstore (the interface layer) so that the test can be
performed easily on other future implementations.

Change-Id: Ie97ebecba82fd5259c5a6380a7f9cf6f48a41f0d