syncbase/vsync: responder discovers initiator
Make the responder learn the name of the initiator during a p2p sync
operation and add the initiator to the SyncGroup memberships of all
common SyncGroups that pass the ACL checks, if the initiator is not
already a member.

This is a temporary workaround until the SyncGroup metadata (which
includes the membership set) starts being synchronized as well.  Until
then, early joiners who are not SyncGroup admins cannot discover later
joiners and start synchronizing with them (initiating to them).  This
issues shows up when testing with 3 or more Syncbases.  This workaround
allows the early joiners to auto-learn of the late joiners when they
receive GetDeltas() requests from them.

This code will be removed when p2p SyncGroup sync is implemented and all
SyncGroup members learn of all SyncGroup mutations (membership updates
as well as SG ACL changes) through the regular p2p sync protocol.

Change-Id: I9cf58cb4291e3443e7519d21630d71827e80ba9f