test/testutil,services: favor NewRandGenerator over InitRandGenerator
As mentioned in the newly added comment in rand.go, using the global singleton
is confusing and can lead to subtle cross-test interference.  It's simpler and
more straightforward to create and use random objects explicitly.

This change makes it possible to do so and changes binarylib and
deviced/impl/internal to the new scheme.

Change-Id: I0193a79e30f9f510e13165e228f65a450211a3db