runtime/internal: hack to support v23 flags in mojo
Mojo services cannot be configured via command-line flags.
We'd like to run Syncbase as a Mojo service, and to configure
v23 in much the same way as we would normally. This CL enables
that by splitting the v23 flag parsing code into "+build mojo"
and "+build !mojo" versions (i.e. dependent on presence of the
"mojo" build tag).

If the "mojo" tag is specified, we parse flags from the env
var V23_MOJO_FLAGS instead of from os.Args.

Change-Id: I6f795c30920b29e5b9f0de58679bbfe91f944b1a