services/device/device: refactor {instance|installation}StateFlag with 'generic'
Since Go doesn't support generics, we define an interface to wrap InstanceState and InstallationState and most of the previously duplicated code now operates on a set of such interface instances.

We come out about even in terms of code, but the benefits of the refactoring
will become more apparent once I add more logic (for implementing exclusion of
states).

Change-Id: I689fbb5b4e15efc6efd90e2a03cd8ba779d14111