services/device/device: Make it an error to operate on the null set.
Sometimes commands like "device update" return success even though they
havn't changed anything. This often is because the device isn't mounted
or something and as a result the "glob" used to run the operation
returns nothing - so the command thinks it has successfully updated 0
objects.

This commit makes it an error to do an operation on 0 objects.
Errors in operating on an object were already being reported.

Change-Id: Ie7302a35b27ac154c1d501bec60eb5f8f7c95070