services/device: Refactor Claim code
This change refactors the claim code such that it can re-used outside of
deviced, and adds a new server that implements Claimable and nothing
else.

The new "claimable" server uses --v23.permissions.* to control who has
access to claim it (instead of a claim token). This approach is more
practical since we don't have to find the claim token. We need to pass
the blessing-root with the --blessing-root flag for this to work.

Change-Id: I124511f872cc5c08fcfab34aafc6bf44612d74ff