v.io/syncbase/x/ref/services/syncbase/localblobstore/blobmap: rename chunkmap to blobmap
This CL renames v.io/syncbase/x/ref/services/syncbase/localblobstore/chunkmap
to v.io/syncbase/x/ref/services/syncbase/localblobstore/blobmap

This is in preparation for storing more data about blobs in the underlying leveldb.
The comments will change to reflect that in the next CL.

This was put in a CL by itself to ease reviewing for the automated parts of the change.

The CL was generated as follows:

	git mv x/ref/services/syncbase/localblobstore/chunkmap x/ref/services/syncbase/localblobstore/blobmap
	git mv x/ref/services/syncbase/localblobstore/blobmap/chunkmap.go x/ref/services/syncbase/localblobstore/blobmap/blobmap.go
	git mv x/ref/services/syncbase/localblobstore/blobmap/chunkmap.go x/ref/services/syncbase/localblobstore/blobmap/blobmap_test.go

	for x in \
			x/ref/services/syncbase/localblobstore/blobmap/blobmap.go \
			x/ref/services/syncbase/localblobstore/blobmap/blobmap_test.go \
			x/ref/services/syncbase/localblobstore/fs_cablobstore/fs_cablobstore.go; do
		sed 's/ChunkMap/BlobMap/g; s/chunkmap/blobmap/g; s/\<cm\>/bm/g' < $x | gofmt > /tmp/x
		cp /tmp/x $x
	done

Change-Id: I66151ea04e0252253b458c65362cc3f01c5284eb