Mojo: Parse command line flags from a variable set at build-time.
We were parsing flags from an environment variable, but that won't work
on Android.

This approach is very flexible.  Any time we build a new binary, we can
determine what flags to run with.

I considered parsing flags from the url parameters, but that doesn't
seem right because the client should not dictate what settings the
service runs with.  It also had implications for caching.

I also considered fetching a config file via HTTP and parsing the flags
from that, but that didn't work because there's no way to tell where the
http server will be running at this point in the code.

Change-Id: I7c438cdc2d134bf3a509d18b8b152c5064e2329c