store: fix a bug in the stream
It turned out that if you cancel a stream that has reached its end
causes the stream to panic. This bug appeared in prefix ACL checking
when we do a scan, get one element from the stream and then cancel
the stream to perform a lower_bound operation on the store.

Change-Id: I5be32e6d69d6498752ef404055c1bbbff0031398