x/ref/services/device/deviced/internal/impl: changes for new socket-based agent
Add support to device manager to work with the new unix doman sockets agent
scheme.

The implementation creates a socket file each time we need to use the app's
principal (using ServePrincipal) and then deletes the socket file (using
StopServing) when no longer needed.  Besides freeing resources in the agent,
this strategy ensures that if the agent goes down and comes back up, we don't
attempt to use a stale socket file (the agent does not yet have support for
re-creating sockets when it restarts). This makes the implementation a bit more
complex since we have to track when the app is not running and stop serving the
socket.

Change-Id: Idd0f94a5f80bf35201758e8b2c43533c738c0ea2