v.io/x/ref/test/benchmark: update to be compatible with go1.5
go 1.5 now defaults to using all of the available CPUs for tests
rather than just one. This means that test/benchmark names will
always have -#cpus suffic when run on a multiproc machine. The
util_test.go tests are not following this convention and this
change modifies them so that they do.

Change-Id: Ic39a8b4d7275c4cb91b8d3e21b620b31102b7824