syncbase: syncQL: support varargs in functions, functions cleanup
StrCat() now takes a variable number of args (but requires at
least two).

In addition, nearly all of the custom arg checking functions
have been deleted.  Writing custom arg checks is usually not
required.  That is, if the only check is to see if the args
match the types specified in the function definitions, nil
can be used for the check function as arg type checking is
now part of the standard check.

Note: There were some wrong types specified for some of
the date functions. Now that the arg types specified are being
used to check the types, they were uncovered and fixed.

Change-Id: I062ea666302a5cd049478f5f3b864bf4c132ff4c