runtime/internal/lib/xwebsocket: Implement ws and wsh protocols for flow.RegisterProtocol.
This change usees websocket framing when possible instead of double
framing connections on websocket connections.
The bulk of the code is the same with minor modifications, switching
from net.conn to flow.MsgReadWriterCloser and net.Listener to
flow.MsgListener, etc.
I also rewrote ws_test to try to make it simpler.

The new code is in an xwebsocket package so that we can remove the
old code more easily once we complete the transition.

Change-Id: I1d7cd8aae89049845acd5a0e6c2dfbac9a91b155