runtime/internal/rpc: Add initial xserver, based on v23/flow.
I started with Matt's xserver.go, which started as a copy of
server.go, deleting stuff that was obviously not necessary, and
changing the code to be based on the flow package.  The code
compiles, but no testing has been performed, and there are TODOs
littered throughout for stuff that we'll need to fix for sure.

Patchset 0 represents the raw copy from server.go to xserver.go,
and patchset 1 represents my edits.

Change-Id: I3a7558c582a425d5b974258dfb5ce96d5807d185