services/device/deviced/internal/impl: fix restart policy and test
Change the logic in the restart policy implementation to match the
documentation: we should reset the restart time window with each restart.

Add a test case in the unit test to capture this behavior; also fix the unit
test comparison logic to correctly evaluate the RestartWindowBegan field (it was
always passing before).  As a consequence of fixing the comparison logic, a test
case needed updating.

Unrelated: add the app instance version's envelope in the Debug output for
convenience and add a couple todos for some issues noticed.

Change-Id: Idb6072354ad7f3920d9dd07fe30203a3596d6f43