services/device/deviced/internal/impl: add stats on app runs and restarts
We want to be able to track how often the device manager restarts individual
apps (collectively and broken down by instance).  Same for how often apps get
run as a result of the Run() rpc.

Unrelated: added a TODO for a crack that apps may fall through depending on when
they die.

Change-Id: I45ee82fc5aebb09733f30c027cc026620dc60995