ref/test/goroutines: Deflake the goroutines test.
Previously I was asserting that the stack of the dumped goroutine should be
exactly X when I knew the depth of the stack.  However it is possible to catch
the stack where it is still inside a wg.Done call for example.  In these
cases the stack could be longer.

Change-Id: I4e7d2b4842c1321077068e29366c0ab2aad44509