lib/security: change ordering of blessing patterns in DebugString to favor ...
This change ensures that the '...' (AllPrincipals) blessing pattern entries
appear first when dumping the blessing store; with all the other patterns
sorted alphabetically afterwards.

There's no strong philosophical argument in favor of this, other than the
convenience of seeing the entry corresponding to 'all principals' in a
recognizable position.  My self-serving impetus to make this change is
v.io/i/680.

Change-Id: I0d653d9c2d767b6aa3d4d22e21af4ddc2b7b7132