syncbase: fix most depcop issues
The remaining "issue" is that our client libs depend on vlog/llog.
It's not clear to me yet whether/how we should avoid this dependency,
so for now I've added a TODO to our .godepcop file.

As part of this change, I ended up cleaning up the organization of
the query package components:
- Placed all query interfaces and impl under a top-level "query" pkg
  (since the design intent is to have this pkg be standalone)
- Moved nosql/internal to nosql/query/internal (see above)
- Made it so that the query pkg does not depend on the nosql pkg
  (again, since that was the intent)
- Moved around some tests to follow idiomatic Go test pkg layout
- Made server-side Database use query.ResultStream, rather than
  defining a redundant ResultStream interface. (Note, this is
  required by the exec.Exec() API - if the interfaces had been
  different, the code would not have compiled.)
- Various other small changes along these lines

MultiPart: 2/2
Change-Id: I8e8b543cc8ddc194ca62d26bd187ef75e7a5243c