namespace: NoResolve is dead, long live Preresolved.  Unlike NoResolve, Preresolved can contain a mount entry to use.
Mounttable access has been simplified.  Instead of trying to apply
all modifying RPCs to all instances of a mount table, we
assume that the instances ensure consistency so that we only have
to reach one.  This reduces an RPC to a mount table to being the
same as an RPC to anything else.  This was the main reason for the
Preresolved change, i.e., to pass multiple servers to a call when
we have preresolved to a mount table.

MultiPart: 2/2
Change-Id: Ie93f6145083e03d0319501d4098860f428adbef9