services/device/device: have publish create new envelope versions
Instead of always using version "0", publish now uploads envelopes at a
new version each time. The version is based on the timestamp.

The envelope from the latest existing version is transfered over to the
new version (keeping with the flag/package preservation behavior of the
old publish command).

Prod and staging need to be updated to fetch updated envelopes from
unversioned object names before checking this in.

MultiPart: 1/2

Change-Id: Id1e059f9bb11c58ea61a9878698b182d29da31b2