ref/runtime/internal/rpc/stream/manager: Fix test hang.
The problem is that the test opens a net.Conn to a stream manager
(just using net.Dial) and then never writes on it.
When we start trying to accept a VIF we have no way of timing out
or cancelling that operation.  We should fix that, but I only
want to deal with it in the new implementation, so for now
I just close the net.Conn which allows everything to shut down.

Change-Id: Ic683ed647ed5475b8287b06d2755edcf10beacea