runtime/internal/rt: remove flakiness from two test cases
TestRemoteStop and TestRemoteForceStop need to wait for the child app to
finish registering a stop listener (m.WaitForStop(ctx, ch)) before
sending a Stop or ForceStop RPC, respectively. Otherwise, the app will
exit with UnhandledStopExitCode.

In practice, we only observed flakiness in TestRemoteStop, since
UnhandledStopExitCode equals ForceStopExitCode which was masking the
flakiness in TestRemoteForceStop.  To make the test more useful and
distinguish between UnhandledStopExitCode and ForceStopExitCode, this CL
also makes it possible to set these two exit codes to different values
for testing purposes (which unearthed another bug in TestForceStop where
we were checking for the wrong exit code).

Since we expect the flakiness in TestRemoteStop to go away after this
CL, we re-enable the test.

MultiPart: 1/2

Change-Id: I99501af3a82d76c00e433485180c36eaf865e1e7