services/internal/binarylib: redo the way we handle duplicate Create()s
This change undoes most of v.io/c/14331 (but keeps the refactoring done
there).  The problem with the solution in v.io/c/14331 is that
specifying methods where we do and don't inherit the root permissions
quickly became clunky, and produced undesired
results (e.g. https://github.com/vanadium/issues/issues/679 or inability
to call Set/GetPermissions on inner nodes even for admin).

This change instead solves the problem that v.io/c/14331 was meant to
fix (namely, competing duplicate Create()s) by making Create atomic.
Specifically, it introduces a SetIfAbsent method on the perms store that
atomically checks if the permissions already exist (meaning that the
binary has already been Created elsewhere) and only sets them if they
don't exist.

While at it, fix the Delete RPC method implementation to also remove the
acls, not just the binary data.

Another (unrelated) fix in this CL is to prevent creating binaries on
the root of the binary server, which offered a trivial way to take over
as admin of the server.

Change-Id: I839db5b8cb94b5933518d7b15d7373330dad32b9