syncbase/nosql: store index for prefix ACLs in a separate key space
This change is required for sync and it solves the following problem:
suppose there are prefix ACLs at F and Foo and there is a sync group
at Foo. Prefix ACL for Foo has a parent F in the local store,
but it could be different for a peer (that also had ACL and Fo).
So we want to keep index for prefix ACLs in a separate key space
and we do not want to sync it. We also need to update the index
every time we set/delete prefix ACLs via sync.

MultiPart: 1/2

Change-Id: Iedf6c004f07eaf73e46b048a3d25a650b88a1415