runtime/internal/gce: Hack for AWS public IP addresses
This is just a quick hack to make sure that services running under
AWS get the correct public IP address to use when publishing
themselves. The part left uncompleted by this change is that the
GCE profile (i.e. all associated files and functions) should really
be renamed to reflect the fact that the profile is handling NAT'ed
cloud VMs, rather than just GCE.

Change-Id: Iff722843983e1fc228cf27c724ca9d7f3120adaa