TBR flow: Remove the events system from conn and replace with simpler mechanism.
It was too dificult to use the events to implement lame ducking.
The problem is that I couldn't detect the transitions I needed (lame duck
is acked, conn is closed) without keeping history.  The code was getting
confused becuase sometimes it would get an event which was really
telling us that the remote end was entering lame duck, but we would
interpret that as an additional local-lame-duck transition.

Hopefully the new code is easier to follow.

Change-Id: I6fdd469c880cd353e43e57d972c267549a453c49