services/application: some fixes to errors returned when there are no envelopes
This CL fixes two distinct problems:

1. When we run Match against a name for which there are no
envelopes, it should return ErrNoExist instead of ErrInvalidSuffix.
This matters for the device publish tool, which treats ErrNoExist to
mean that there are no envelopes yet uploaded for the given
application (whereas if it sees ErrInvalidSuffix, it complains and
exits)

2. When we run Glob or TidyNow against the root of the application
server and there are no applications on the server, it is failing with
the confusing:
applicationd:"".[__Glob|TidyNow]: Does not exist: applications
The reason has to do with what
i.store.BindObject("/applications").Children() returns when there are no
actual app objects under /applications in the server's internal
store. The fix is to properly handle this case.

Unrelated to the above, updating the documentation for application
remove to mention the option of using '*' for all profiles.

Change-Id: I8e52e6452f40bdcf7c5dba89a37f1098c3ecfde7