services/device/deviced/internal/impl/utiltest: fix WaitForMount retry logic
As Matt observed, the retry logic in WaitForMount doesn't work when the
mounttable resolution returns an error (such as when the name doesn't
exist) because Resolve is wired to fail in that case.

To fix this, we stop using Resolve inside WaitForMount and instead call
the namespace's Resolve method directly.

Change-Id: Ib463bdf16e6782527d56f2850e7f7ec55db40de6