services/profiled: Make tests pass under the new RPC system.
Before we created two servers each built on the context of the last.
We then cancel the first server and expect to use the second.  However
the second server (and it's attached client) have their context canceled
at the same time the first server is closed.

The fix is to simply not keep the contexts returned by the new servers.

Change-Id: I1df8f2ebef38abf17c8943a05c5dcdc90256ee0f