service/cluster: Add Cluster and Pod Agents
This change implements the Cluster Agent described in
https://docs.google.com/document/d/1fHZC9F3lyFo7v4bl5o9Bs1uNjE8-TvBD5PCl36UFlyg/view#heading=h.gvse1ra7h9wn

The Cluster Agent keeps a list of Secret Keys, one for each Service,
and Blessings associated with them. It issues new Blessings to whoever
presents a valid Secret Key. The new Blessings are extensions of the
Blessings associated with the Secret Key.

The Pod Agent implements the normal Agent interface (i.e. with
V23_AGENT_PATH and a unix socket), but internally talks to the Cluster
Agent to get its blessings. The Principal served by this Agent is
read-only.

Change-Id: Id9976603689b479db40e8afd35b3e67e9b1f4c91