services/binary/binary: make download install the binary after downloading it
This matters for packages that are archived directories.  Based on the
value of the new "--install" flag, download will fetch the binary and
media info under a temp location and extract to the specified
destination, or just download the binary and media info as the specified
destination.  For installation, if the binary is a regular file, a hard
link (or copy) is created to the temp location. If it's an archive, it
gets extracted.

While at it, we also change the tar extraction logic in the packages
library to work with intermediary directories.

Change-Id: I58b1681a5cdc824bd4557706538d07a313ebec82