ref: Ensure endpoint used for StartCall is used for conn.Dial.
Scenario:
User uses a endpoint without a blessing pattern to remote end, thus
making an unauthorized connection to the remote end.
Then client then decides to specify a endpoint with a blessing pattern.
We need to ensure that this new endpoint with the blessing pattern is
the one used to authorize the server during conn.Dial.

This means that flows on the same conn may have a different RemoteEndpoints,
thus the API changes.

Change-Id: Ia8c8a638d9c8ef153a135f6f3e77a7880e22a0ec
MultiPart: 2/2