TBR: test/goroutines: Fix TestFormat.
The issue was that we were comparing the Parsed goroutines to the
actual runtime stack, but we have since added the ability to ignore
goroutines during parse. This caused this test to be flaky.

The fix implemented here is to add a parameter to Parse that allows
not ignoring the goroutines.

Change-Id: I259b1c0320509310a0b5edacbb2344808e165b07