services/cluster: Add vkube command.
This is a prototype that implements the basic functionality of a cluster
manager. The vkube command can be used to:
 - start / stop the cluster agent.
 - start / update / stop vanadium applications.
 - build docker images for cluster & pod agents.

When vanadium application are started, vkube creates a new secret key on
the cluster agent, and a Secret object on Kubernetes.

This Secret object is then added to the user-provided
replication-controller config, along with a pod agent container and all
the hooks necessary for the vanadium app to talk to the pod agent.

Tested manually against a Kubernetes cluster. Still need to add a lot
tests.

Change-Id: I8fa6c00e8011c996eefe074184a2d8bdf1ba087f