lib/discovery: Hack to get the discovery factories to work on iOS.
The current ble plugin implementation doesn't work on iOS because the
paypal gatt libraries don't work on iOS because xrpc/xrpc.h is not
available when building for iOS, seemingly because xrpc is not part of
the iOS SDK.
(gomobile --target=ios build ...
fails)

The concrete use case right now for iOS builds is the volley app:
https://github.com/monopole/volley
and trying to get mDNS based discovery working there.

For now, this commit just disables BLE discovery on iOS completely.
Making it work is left as an exercise to folks in the future.

Change-Id: I85ecb1381bbe20a9bb2b41b8594c1cd409e503e9