services/device/dmrun: ensure workdir is there when vcloud gets built
Move the creation of the working dir before the handleFlags call, since
inside handleFlags we may build vcloud and put it in the dir.  This
should fix the issue of having vcloud show up in the current dir.  Also,
move the check for no args before doing most of the work in handleFlags.

Change-Id: Ib0a971ddd6bc1deb59e0af44f520e46aaa3240fd