services/device/mgmt_v23_test: de-flake: wait for associate command to complete
We should wait for the associate command to complete before going ahead
and checking the result of the association immediately below.

This can cause occasional flakiness if the associate command has not yet
had a chance to do its job by the time the next command (which lists the
associations) executes.

Should fix https://github.com/vanadium/issues/issues/608

Change-Id: If74637911cb13f1110f1dddeea31b4be4b6122b1