ref: Implement roaming.
The flow manager now handles roaming.
Changes:
(1) The server is notified to requery ListeningEndpoints when a chan
returned from ListeningEndpoints is closed.
(2) Roaming logic has been refactored into a library.
(3) WatchNetwork and UnwatchNetwork have been deprecated and replaced
by status.Valid that is a channel that is closed when the
Status needs to be requeried by callers.

MultiPart: 2/2

Change-Id: I5e34876205e1017ae8be00ca5bb731fa0a94a5a2