services/cluster/vkube: Add integration test
This change adds an integration test for vkube. It is not currently
possible to run on it jenkins.

The test needs:
- access to docker to build docker images.
- access to gcloud and kubectl, gcloud must be pre-initialized.
- access to a pre-existing kubernetes cluster.

It takes ~10 minutes to run, and can easily exceed the default test
timeout. I would recommend running it with -test.timeout=20m.

It is OK to run multiple instances of the test concurrently. Each
instance creates its own bucket to store docker images, and its own
namespace on the cluster. Both are deleted when the test ends.

Change-Id: I599eaaf04967cbef1023f6246150709f76d81ee1