services/syncbase: add flag to disable neighborhood and other small fixes
I found myself needing to turn off publishing in the neighborhood -- the
reason was having two syncbased servers on the same machine avertising
under the same name. I expect this particular issue will get resolved,
but it still seems useful to offer a way to disable publishing in the
neighborhood for other reasons.

Other changes:

- log the actual error in syncWithPeer()'s GetTime instead of a generic
  error that lacks any information

- remove the extra 'call' argument from NewService() and New() -- it's
  neither used nor passed in.

MultiPart: 1/2

Change-Id: I5ba8109cdee0d585da7c65fbce6b746771f7dd65