dmrun: show a good error message if ssh fails
Before this change, if the dmrun tool was used in the ssh mode
and the ssh command failed, then the tool just wrote:
"working dir /tmp/dmrun already exists on target.
Please clean up any previous dmrun instance."
This was the first ssh command and it could fail just because of
bad arguments passed to the ssh command.
And now we fix that.

Change-Id: Icf0124548930df418e1f41d23fac45c38ed6fdf5