"x/ref": provide a reason with blessing flow close error
In the new RPC framework, whenever a receive on a blessings flow
fails, the flow is closed and a "blessings flow closed" error
is returned to the user at the local end. This is somewhat
incomplete as the reason for why the flow was closed is not revealed
by the error.

This CL tries to fix this by nesting the receive error inside
the blessings flow closed error.

Change-Id: I9f304b32f9a86d98ae8652341963ff12dc2c85f5