mgmt: Explicitly provide the blessing for the app cycle manager.
With this patch, the device manager explicitly provides the blessing
that the appcycle manager should use to communicate with it. Since
the appcycle manager only communicates with the device manager, it
can use that blessing exclusively and doesn't have to pollute the
blessing store of the application, and doesn't have to use any other
blessings of the application.

To enable smooth rollout, a bool is provided that maintains
compatibility with old behavior. I intend to remove that bool
and code dependent on it in 1-2 weeks.

Change-Id: I2a88764136870ebd264c695eecdb176a0e4d76be