services/xproxy: Add publishing to proxy and first iteration of xproxyd binary.
Done in this change:
(1) Moved ref/runtime/internal/lib/publisher to ref/lib/publisher so
    that xproxy can use it. This involved fixing the publisher tests
    to use v23.Init.
(2) Proxy publishes itself under a name.
(3) Proxy listens through other proxies if the Proxy field of the
    listenSpec is specified.
(4) xproxyd binary is written.

TODO still (including things realized while writing this):
(1) The name of a proxy should be reresolved when reconnecting to it.
(2) The proxy should use a authorizer to authorize servers trying to
    connect to it.
(3) The proxy should send update messages to servers as its endpoints
    change.

Change-Id: I19284a241df6e17cabe9b9b0693d38005784f95c