services/syncbase: Obtain blessings from the Mojo authentication service.
When syncbase is running as a mojo service (and under android for now),
it will obtain the blessings to use from the Mojo authentication
service.

This is just an intermediate step which allows syncbase instances on
different devices (running mojo) to establish blessings (identities)
that they can use to authorize each other. Mojo applications using the
syncbase service themselves are not authenticated - so all applications
on the device have the same level of access to data in the local
syncbase.

Details are dumped into a large comment at the top of mojo_blessings.go.

Change-Id: I40e730c923fc3e95b97ab1c410a2aef7096d71fa