runtime/factories: Remove the static factory.
There is no advantage to using the static factory over the roaming one.
When the ListenSpec is provided with a specific address (such as
IP:port), then the roaming doesn't take effect anyway and in terms of
code being linked in, there is practically no difference between the
two.

Remove it and there will be no confusion as to which one to use.

MultiPart: 1/2

Change-Id: I241d3cd535be0f8fa2595ac617a002ca8372a491