services/identity: Use local endpoint for macaroon service
With this change the Macaroon that is created by the OAuth handler will
be validated by the MacaroonService running on the same instance.

This is useful when the identityd service is replicated and all the
replicas publish themselves under the same name.

For the Discharger service, all the replicas are interchangeable (as
long as they use the same key). For the Macaroon service, the replicas
are NOT interchangeable. Each Macaroon is created with an ephemeral key
and must be validated by the replica that created it.

Change-Id: I8e0023e8a7b34f6cd9ee3e6b5bd4f07752cc7107