rpc: Ensure that if a roaming channel is already closed we don't hang trying to read from it.
The channel was closed from two places, either if the publisher
finishes, or if we CloseFork because the receiver is finished.
In the latter case the reciever must drain the channel.  If the
sender closed the channel just as the reciever closed fork the
reciever might read forever from the closed channel trying to drain it.

Change-Id: I99dda1153208f367e0f6c8419026e8f45effd897